real_estate_ai/requirements.txt
fastapi
uvicorn
python-dotenv
pymongo
pydantic
google-genai
playwright